opsui/traits: relocate traitCells/Cards into traits/'trait' folder and update component imports
Jira: N/A
