+++ ui/conductor/src/routes/devices/components/DeviceCell.vue
@@ -106,16 +106,16 @@ import WithMeter from '@/routes/devices/components/renderless/WithMeter.vue';
 import WithOccupancy from '@/routes/devices/components/renderless/WithOccupancy.vue';
 import WithOpenClosed from '@/routes/devices/components/renderless/WithOpenClosed.vue';
 import WithStatus from '@/routes/devices/components/renderless/WithStatus.vue';
-import AccessAttemptCell from '@/routes/devices/components/trait-cells/AccessAttemptCell.vue';
-import AirTemperatureCell from '@/routes/devices/components/trait-cells/AirTemperatureCell.vue';
-import ElectricDemandCell from '@/routes/devices/components/trait-cells/ElectricDemandCell.vue';
-import EmergencyCell from '@/routes/devices/components/trait-cells/EmergencyCell.vue';
-import EnterLeaveEventCell from '@/routes/devices/components/trait-cells/EnterLeaveEventCell.vue';
-import LightCell from '@/routes/devices/components/trait-cells/LightCell.vue';
-import MeterCell from '@/routes/devices/components/trait-cells/MeterCell.vue';
-import OccupancyCell from '@/routes/devices/components/trait-cells/OccupancyCell.vue';
-import OpenClosedCell from '@/routes/devices/components/trait-cells/OpenClosedCell.vue';
-import StatusLogCell from '@/routes/devices/components/trait-cells/StatusLogCell.vue';
+import AccessAttemptCell from '@/traits/access/AccessAttemptCell.vue';
+import AirTemperatureCell from '@/traits/airTemperature/AirTemperatureCell.vue';
+import ElectricDemandCell from '@/traits/electricDemand/ElectricDemandCell.vue';
+import EmergencyCell from '@/traits/emergency/EmergencyCell.vue';
+import EnterLeaveEventCell from '@/traits/enterLeave/EnterLeaveEventCell.vue';
+import LightCell from '@/traits/lighting/LightCell.vue';
+import MeterCell from '@/traits/meter/MeterCell.vue';
+import OccupancyCell from '@/traits/occupancy/OccupancyCell.vue';
+import OpenClosedCell from '@/traits/openClosed/OpenClosedCell.vue';
+import StatusLogCell from '@/traits/status/StatusLogCell.vue';
 import {hasTrait} from '@/util/devices';
 import {computed} from 'vue';
