+++ ui/conductor/src/traits/airTemperature/AirTemperatureCard.vue
