+++ ui/conductor/src/traits/airTemperature/AirTemperatureCell.vue
