+++ ui/conductor/src/traits/deviceInfo/DeviceInfoCard.vue
