+++ ui/conductor/src/traits/electricDemand/ElectricDemandCell.vue
