+++ ui/conductor/src/traits/emergency/EmergencyCell.vue
