+++ ui/conductor/src/traits/emergency/EmergencyLight.vue
