+++ ui/conductor/src/traits/mode/ModeCard.vue
