+++ ui/conductor/src/traits/occupancy/OccupancyCard.vue
