+++ ui/conductor/src/traits/occupancy/OccupancyCell.vue
