+++ ui/conductor/src/traits/openClosed/OpenClosedCard.vue
