+++ ui/conductor/src/traits/openClosed/OpenClosedCell.vue
