+++ ui/conductor/src/traits/status/StatusLogCard.vue
