+++ ui/conductor/src/traits/status/StatusLogCell.vue
