+++ ui/conductor/src/traits/udmi/UdmiCard.vue
